Broke apart maths.positional into separate submodules, under the new "numbersys" maths module
